requirements.txt
pandas==2.0.3
tabula-py==2.8.0
pdfplumber==0.10.2
openpyxl==3.1.2
Pillow==10.0.0
pyinstaller==5.13.0
numpy==1.26.4